pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hydrothermal-relithiation-workflows"
version = "0.1.0"
description = "Computational workflows for hydrothermal cathode relithiation experiments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "numpy>=2.0,<2.4",
  "scikit-learn>=1.6.1",
  "matplotlib",
  "tqdm",
  "maml",
  "ase",
  "pymatgen",
  "hiphive",
  "fairchem-core>=2.15.0,<3",
  "torch-sim-atomistic",
  "matgl>=2.0.0",
  "torch~=2.8.0",
]

[project.optional-dependencies]
matgl-dgl = [
  "dgl>=2.2.0,<2.5",
]
dev = [
  "pytest",
  "ruff",
  "mypy",
]

[project.scripts]
hrw = "hydrorelith.cli:main"
hrw-electrode-generate = "hydrorelith.pipelines.electrode_structure_generation:main"
hrw-electrolyte-generate = "hydrorelith.pipelines.electrolyte_structure_generation:main"
hrw-scf-parallel-benchmark = "hydrorelith.pipelines.scf_parallelization_benchmark:main"
hrw-uma-torchsim-screen = "hydrorelith.pipelines.uma_torchsim_screen:main"
hrw-uma-torchsim-analyze = "hydrorelith.pipelines.uma_torchsim_screen:main_analyze"
hrw-uma-torchsim-export2pt = "hydrorelith.pipelines.uma_torchsim_screen:main_export2pt"
hrw-uma-torchsim-manifest-template = "hydrorelith.io.structure_manifest:main_template"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
